task-service/requirements.txt
alembic==1.0.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
consulate==0.6.0
dnspython==1.15.0
fl-model==0.0.1
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.3
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.19.1
six==1.11.0
slerp-py==0.0.19
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
